+++ src/test/java/com/epam/reportportal/utils/markdown/MarkdownUtilsTest.java
@@ -17,11 +17,9 @@ package com.epam.reportportal.utils.markdown;
 
 import org.junit.jupiter.api.Test;
 
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
 
-import static com.epam.reportportal.utils.markdown.MarkdownUtils.asCode;
-import static com.epam.reportportal.utils.markdown.MarkdownUtils.formatDataTable;
+import static com.epam.reportportal.utils.markdown.MarkdownUtils.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 
@@ -29,13 +27,16 @@ import static org.hamcrest.Matchers.equalTo;
  * @author Andrei Varabyeu
  */
 public class MarkdownUtilsTest {
+	//@formatter:off
+	public static final String ONE_ROW_EXPECTED_TABLE =
+			TABLE_INDENT + "|\u00A0var_a\u00A0|\u00A0var_b\u00A0|\u00A0result\u00A0|\n"
+					+ TABLE_INDENT + "|-------|-------|--------|\n"
+					+ TABLE_INDENT + "|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A04\u00A0\u00A0\u00A0\u00A0|";
 
-	public static final String ONE_ROW_EXPECTED_TABLE = "\u00A0\u00A0\u00A0\u00A0|\u00A0var_a\u00A0|\u00A0var_b\u00A0|\u00A0result\u00A0|\n"
-			+ "\u00A0\u00A0\u00A0\u00A0|-------|-------|--------|\n"
-			+ "\u00A0\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A04\u00A0\u00A0\u00A0\u00A0|";
-
-	public static final String TWO_ROWS_EXPECTED_TABLE = ONE_ROW_EXPECTED_TABLE + "\n"
-		+ "\u00A0\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A01\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A03\u00A0\u00A0\u00A0\u00A0|";
+	public static final String TWO_ROWS_EXPECTED_TABLE =
+			ONE_ROW_EXPECTED_TABLE
+					+ "\n" + TABLE_INDENT + "|\u00A0\u00A0\u00A01\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A03\u00A0\u00A0\u00A0\u00A0|";
+	//@formatter:on
 
 	@Test
 	public void asMarkdown() {
@@ -47,16 +48,68 @@ public class MarkdownUtilsTest {
 		assertThat("Incorrect markdown prefix", asCode("groovy", "hello"), equalTo("!!!MARKDOWN_MODE!!!```groovy\nhello\n```"));
 	}
 
-
 	@Test
 	public void test_format_data_table() {
-		List<List<String>> table =
-				Arrays.asList(
-						Arrays.asList("var_a", "var_b", "result"),
-						Arrays.asList("2", "2", "4"),
-						Arrays.asList("1", "2", "3")
-				);
+		List<List<String>> table = Arrays.asList(Arrays.asList("var_a", "var_b", "result"),
+				Arrays.asList("2", "2", "4"),
+				Arrays.asList("1", "2", "3")
+		);
 
 		assertThat(formatDataTable(table), equalTo(TWO_ROWS_EXPECTED_TABLE));
+	}
+
+	public static final String BIG_COLUMN_VALUE = "4CZtyV3qsjVX08vBKu5YpvY2ckoFxLUombHEj1yf4uaBmrSGTzcXvlfba52HtUGLm56a8Vx4fBa0onEjlXY";
+	public static final String TRUNCATED_COLUMN_VALUE = "4CZtyV3qsjVX08vBKu5YpvY2ckoFxLUombHEj1yf4uaBmrSGTzcXvlfba52...";
+
+	//@formatter:off
+	public static final String ONE_ROW_LONG_EXPECTED_TABLE =
+			TABLE_INDENT + "|\u00A0var_a\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0var_b\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|\u00A0result\u00A0|\n"
+					+ TABLE_INDENT + "|-------|----------------------------------------------------------------|--------|\n"
+					+ TABLE_INDENT + "|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A04\u00A0\u00A0\u00A0\u00A0|";
+
+	public static final String TWO_ROWS_LONG_EXPECTED_TABLE =
+			ONE_ROW_LONG_EXPECTED_TABLE + "\n"
+					+ TABLE_INDENT + "|\u00A0\u00A0\u00A01\u00A0\u00A0\u00A0|\u00A0" + TRUNCATED_COLUMN_VALUE + "\u00A0|\u00A0\u00A0\u00A03\u00A0\u00A0\u00A0\u00A0|";
+	//@formatter:on
+
+	@Test
+	public void test_format_data_table_one_big_col() {
+		List<List<String>> table = Arrays.asList(Arrays.asList("var_a", "var_b", "result"),
+				Arrays.asList("2", "2", "4"),
+				Arrays.asList("1", BIG_COLUMN_VALUE, "3")
+		);
+		assertThat(formatDataTable(table), equalTo(TWO_ROWS_LONG_EXPECTED_TABLE));
+	}
+
+	public static final String SECOND_TRUNCATED_COLUMN_VALUE = "4CZtyV3qsjVX08vBKu5YpvY2ckoFxLU...";
+
+	//@formatter:off
+	public static final String ONE_ROW_LONG_EXPECTED_TABLE_TWO =
+			TABLE_INDENT + "|\u00A0var_a\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0var_b\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0result\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|\n"
+					+ TABLE_INDENT + "|-------|------------------------------------|------------------------------------|\n"
+					+ TABLE_INDENT + "|\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A02\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|\u00A0" + SECOND_TRUNCATED_COLUMN_VALUE + "\u00A0|";
+
+	public static final String TWO_ROWS_LONG_EXPECTED_TABLE_TWO =
+			ONE_ROW_LONG_EXPECTED_TABLE_TWO + "\n"
+					+ TABLE_INDENT + "|\u00A0\u00A0\u00A01\u00A0\u00A0\u00A0|\u00A0" + SECOND_TRUNCATED_COLUMN_VALUE + "\u00A0|\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A03\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0|";
+	//@formatter:on
+
+	@Test
+	public void test_format_data_table_two_big_col() {
+		List<List<String>> table = Arrays.asList(Arrays.asList("var_a", "var_b", "result"),
+				Arrays.asList("2", "2", BIG_COLUMN_VALUE),
+				Arrays.asList("1", BIG_COLUMN_VALUE, "3")
+		);
+		assertThat(formatDataTable(table), equalTo(TWO_ROWS_LONG_EXPECTED_TABLE_TWO));
+	}
+
+	@Test
+	public void test_format_data_table_map() {
+		Map<String, String> table = new LinkedHashMap<String, String>(){{
+			put("var_a", "2");
+			put("var_b", "2");
+			put("result", "4");
+		}};
+		assertThat(formatDataTable(table), equalTo(ONE_ROW_EXPECTED_TABLE));
 	}
 }